refactor(RacingReport) : RacingReport 리팩토링 전체 코드 적용 및 코드 정리
- RacingReportRefactor.java 제거 하고 RacingReport.java 로 적용
- RacingReport 변경된 코드로 프로젝트 적용
- 코드 리팩토릭 진행
